Made the login page and manage account section look prettier
