+++ client/components/layout.js
@@ -5,9 +5,6 @@ import styles from './layout.module.css';
 import Header from './header.js';
 import Footer from './footer.js';
 import { AuthorizationProvider } from '../context/AuthorizationContext';
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
-
-const name = 'Nathan';
 
 export default function Layout({ children }) {
     return (<div className={styles.container}>
