+++ client/components/layout.module.css
@@ -3,3 +3,7 @@
     padding: 0 1rem;
     margin: 3rem auto 6rem;
 }
+
+.footerSeparator {
+    margin-top: 2em;
+}
