+++ client/components/userInfo.js
@@ -4,6 +4,7 @@ import API from '../apis/api.js';
 import useAuth from '../hooks/useAuthorization.js';
 import utilStyles from '../styles/utils.module.css';
 import buttonStyles from './buttons.module.css';
+import loginStyles from './loginStyles.module.css';
 
 function UserInfo({ username }) { 
     const { auth, setAuth } = useAuth();
@@ -15,6 +16,7 @@ function UserInfo({ username }) {
     });
 
     const [changePasswordAlert, setChangePasswordAlert] = useState(false);
+    const [changePasswordStatus, setChangePasswordStatus] = useState('');
     const [changePasswordMsg, setChangePasswordMsg] = useState('');
 
     const handlePasswordChange = (e) => {
@@ -22,12 +24,14 @@ function UserInfo({ username }) {
             ...changePasswordDetails,
             [e.target.name]: e.target.value
         });
-        setChangePasswordMsg(' ');
+        setChangePasswordStatus('');
+        setChangePasswordMsg('');
     };
     const changePasswordSubmit = async (e) => {
         e.preventDefault();
         if (changePasswordDetails.newPassword !== changePasswordDetails.confirmPassword) {
-            setChangePasswordMsg("Change password failed: confirm password doesn't match new password");
+            setChangePasswordStatus("Failed");
+            setChangePasswordMsg("confirm password doesn't match new password");
             setChangePasswordAlert(true);
             return null
         }
@@ -37,16 +41,22 @@ function UserInfo({ username }) {
                 oldPassword: changePasswordDetails.oldPassword,
                 newPassword: changePasswordDetails.newPassword
             });
-            setChangePasswordMsg("Password changed");
+            setChangePasswordStatus("Password changed");
+            setChangePasswordDetails({
+                oldPassword: "",
+                newPassword: "",
+                confirmPassword: ""
+            });
         } catch(err) {
+            setChangePasswordStatus("Failed");
             if (!err?.response) {
-                setChangePasswordMsg("Change password failed: no server response");
+                setChangePasswordMsg("no server response");
             } else if (err.response?.status === 403) {
-                setChangePasswordMsg("Change password failed: Current user doesn't exist");
+                setChangePasswordMsg("current user doesn't exist");
             } else if (err.response?.status === 401) {
-                setChangePasswordMsg("Change password failed: incorrect password");
+                setChangePasswordMsg("incorrect password");
             } else {
-                setChangePasswordMsg("Change password failed");
+                setChangePasswordMsg("change password failed");
             }
         };
         setChangePasswordAlert(true);
@@ -59,44 +69,49 @@ function UserInfo({ username }) {
         router.push("/");
     }
     return (
-        <section>
-            <h3>Manage Account</h3>
-            <button onClick={e => signOut(e)} className={buttonStyles.listDelete}>Sign Out</button>
-            <h3>Change Password</h3>
+        <section className={loginStyles.container}>
+            <h3 className={loginStyles.heading}>Change Password</h3>
             <form onSubmit={(e) => changePasswordSubmit(e)}>
-                <label htmlFor="oldPassword">Enter old password: </label>
+                <label htmlFor="oldPassword">Enter old password</label><br/>
                 <input 
                     type="password" 
                     id="oldPassword" 
                     name="oldPassword" 
                     placeholder="old password" 
+                    value={changePasswordDetails.oldPassword}
                     className={utilStyles.textInput}
                     onChange={e => handlePasswordChange(e)}
                     required
                 /><br/>
-                <label htmlFor="newPassword">Enter new password: </label>
+                <label htmlFor="newPassword">Enter new password</label><br/>
                 <input 
                     type="password" 
                     id="newPassword" 
                     name="newPassword" 
                     placeholder="new password"
+                    value={changePasswordDetails.newPassword}
                     className={utilStyles.textInput}
                     onChange={e => handlePasswordChange(e)}
                     required
                 /><br/>
-                <label htmlFor="confirmPassword">Confirm new password: </label>
+                <label htmlFor="confirmPassword">Confirm new password</label><br/>
                 <input 
                     type="password" 
                     id="confirmPassword" 
                     name="confirmPassword" 
                     placeholder="confirm password"
+                    value={changePasswordDetails.confirmPassword}
                     className={utilStyles.textInput}
                     onChange={e => handlePasswordChange(e)}
                     required
                 /><br/>
                 <input type="submit" value="Change Password" className={buttonStyles.listEdit}/>
-                {changePasswordAlert && <p>{changePasswordMsg}</p>}
+                {changePasswordAlert && <>
+                <p className={utilStyles.noMargin}>{changePasswordStatus}</p>
+                    <p className={utilStyles.noMargin}>{changePasswordMsg}</p></>}
             </form>
+            <hr className={loginStyles.containerRule}/>
+            <button onClick={e => signOut(e)} className={buttonStyles.listDelete}>Sign Out</button>
         </section>
     );
 }
