+++ client/package.json
@@ -9,7 +9,6 @@
     "axios": "^1.4.0",
     "next": "^13.3.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-router-dom": "^6.11.2"
+    "react-dom": "^18.2.0"
   }
 }
