+++ client/styles/global.css
@@ -19,7 +19,7 @@ h1, h2, h3 {
 }
 
 a {
-    color: #0070f3;
+    color: #83a598;
     text-decoration: none;
 }
